+++ src/main/java/com/api/management/leave/leavemanagementapi/utils/HourConversion.java
@@ -1,19 +1,21 @@
 package com.api.management.leave.leavemanagementapi.utils;
 
+import java.math.BigDecimal;
+
 public enum HourConversion {
-    ONE(1, .125),
-    TWO(2, .250),
-    THREE(3, .375),
-    FOUR(4, .500),
-    FIVE(5, .625),
-    SIX(6, .750),
-    SEVEN(7, .875),
-    EIGHT(8, 1.000);
+    ONE(1, BigDecimal.valueOf(.125)),
+    TWO(2, BigDecimal.valueOf(.250)),
+    THREE(3, BigDecimal.valueOf(.375)),
+    FOUR(4, BigDecimal.valueOf(.500)),
+    FIVE(5, BigDecimal.valueOf(.625)),
+    SIX(6, BigDecimal.valueOf(.750)),
+    SEVEN(7, BigDecimal.valueOf(.875)),
+    EIGHT(8, BigDecimal.valueOf(1.000));
 
     private final int hour;
-    private final double equivalentDay;
+    private final BigDecimal equivalentDay;
 
-    HourConversion(int hour, double equivalentDay) {
+    HourConversion(int hour, BigDecimal equivalentDay) {
         this.hour = hour;
         this.equivalentDay = equivalentDay;
     }
@@ -22,7 +24,7 @@ public enum HourConversion {
         return hour;
     }
 
-    public double getEquivalentDay() {
+    public BigDecimal getEquivalentDay() {
         return equivalentDay;
     }
